requirements.txt
fastapi 
uvicorn 
python-multipart 
python-jose[cryptography] 
authlib 
python-dotenv 
fastapi-microsoft-identity
